refactor: create social-icon folder and move icons inside to keep all component files organised
